refactor: 컴포넌트 분리 및 useOutsideClick 적용
